requirements.txt
tensorflow==2.9.1
Pillow==9.1.0
numpy==1.23.1
opencv-python-headless==4.6.0.66